PianoTutor patch
This experimental patch provides a full piano tutoring/learning system
exploiting a simple RGB LED strip (a NeoPixel with 144 or 72 LEDs in
1m would work) driven via a separate Arduino program, and a MIDI/USB
capable keyboard.

Two new tutor checkboxes on the main toolbar allow for:
a) enabling the note on and off messages to be reflected on the LEDs,
   so that the keys being played are visually highlighted right on the
   keyboard;
b) enabling the tutor/learning mode, in which for each key sequence
   encountered during playback with non-muted notes, MuseScore turns
   on the corresponding keys on the LED strip, then it waits for them
   to be all pressed by the user on the keyboard, before moving on
   with the playback; MuseScore follows the user key presses in real-time
   during playback.

It is possible to move throughout the score using MuseScore, starting
the playback and/or learning session wherever desired. It is possible
to select a region of the score (typically a few measures), then use
the MuseScore Loop function to keep playing and practising the same
region till the user is confident to move to the next phrase to learn.

In a typical 2-staves piano score, it is possible to mute one hand and
practise only with the other one, then switch them, then play and
practise them together, using the usual mute function available in
MuseScore within the Mixer pane.

Technical details

MuseScore sends to the Arduino program the key numbers whose color has
to change over the serial/USB port (normally /dev/ttyACM0 or ttyACM1),
in 115200 8N1 mode, along with the new color, in the fixed 17-bytes
format:

  k<key>r<red>g<green>b<blue><eol>

where all <numbers> are 3-digit, decimal numbers, e.g.:

  k060r127g000b000\n

turns on a RED light on the 61st LED (with the first LED, having the
key code 0, on the lowest C on a 88-keys piano, this should be the
"middle C"). The sequence

  k060r000g000b000

turns off the corresponding LED.

Note on key sizes: with my old Roland digital piano, each key spans
two LEDs of a 144/m strip, and alignment is not perfect, but
sufficient to not generate confusion among which LED corresponds to
which key. For other keyboards the situation may be different. In
theory, alignment issues should be fixable by changing the "2*"
multiplier in seq.cpp with the actual LEDs/keys ratio.