PianoTutor - lookahead feature
This patch adds a third checkbox to the PianoTutor system, letting
you see on the LEDs not only the next set of keys to press, but
also the sequence coming right after, in a dimmer light. This might
help in getting the fingering right earlier, when tackling a piece
never played before.